+++ 1.21.8/Common/src/main/java/fuzs/effectdescriptions/client/handler/EffectWidgetHandler.java
@@ -4,7 +4,6 @@ import fuzs.effectdescriptions.EffectDescriptions;
 import fuzs.effectdescriptions.client.helper.EffectTooltipSuppliers;
 import fuzs.effectdescriptions.config.ClientConfig;
 import fuzs.puzzleslib.api.client.gui.v2.ScreenHelper;
-import fuzs.puzzleslib.api.client.gui.v2.tooltip.ClientComponentSplitter;
 import fuzs.puzzleslib.api.event.v1.core.EventResult;
 import fuzs.puzzleslib.api.event.v1.data.MutableBoolean;
 import fuzs.puzzleslib.api.event.v1.data.MutableInt;
@@ -12,40 +11,43 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
-import net.minecraft.client.gui.screens.inventory.tooltip.DefaultTooltipPositioner;
 import net.minecraft.network.chat.Component;
-import net.minecraft.util.FormattedCharSequence;
 import net.minecraft.world.effect.MobEffectInstance;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 
 public class EffectWidgetHandler {
     @Nullable
-    private static List<FormattedCharSequence> tooltipLines;
+    private static List<Component> tooltipLines;
 
     public static EventResult onInventoryMobEffects(Screen screen, int availableSpace, MutableBoolean smallWidgets, MutableInt horizontalOffset) {
-        if (!EffectDescriptions.CONFIG.get(ClientConfig.class).widgetTooltips) return EventResult.PASS;
-        if (!(screen instanceof AbstractContainerScreen<?> abstractContainerScreen)) return EventResult.PASS;
+        if (!EffectDescriptions.CONFIG.get(ClientConfig.class).widgetTooltips) {
+            return EventResult.PASS;
+        }
+
+        if (!(screen instanceof AbstractContainerScreen<?> abstractContainerScreen)) {
+            return EventResult.PASS;
+        }
+
         if (isInWidgetsColumn(horizontalOffset.getAsInt(), smallWidgets.getAsBoolean())) {
             MobEffectInstance mobEffect = getMobEffectInRow(abstractContainerScreen);
+
             if (mobEffect != null) {
-                List<Component> list = EffectTooltipSuppliers.getMobEffectComponents(mobEffect);
-                tooltipLines = ClientComponentSplitter.splitTooltipLines(list).toList();
+                // there is no GuiGraphics instance available from the event,
+                // so temporarily store the tooltip and submit it in a later render event
+                tooltipLines = EffectTooltipSuppliers.getMobEffectComponents(mobEffect);
             }
         }
+
         return EventResult.PASS;
     }
 
     public static void onDrawForeground(AbstractContainerScreen<?> screen, GuiGraphics guiGraphics, int mouseX, int mouseY) {
         if (tooltipLines != null) {
-            guiGraphics.setTooltipForNextFrame(screen.getFont(),
-                    tooltipLines,
-                    DefaultTooltipPositioner.INSTANCE,
-                    mouseX,
-                    mouseY,
-                    true);
+            guiGraphics.setTooltipForNextFrame(screen.getFont(), tooltipLines, Optional.empty(), mouseX, mouseY);
             tooltipLines = null;
         }
     }
@@ -66,8 +68,10 @@ public class EffectWidgetHandler {
             if (mouseY >= posY && mouseY <= posY + verticalOffset) {
                 hoveredMobEffect = mobEffect;
             }
+
             posY += verticalOffset;
         }
+
         return hoveredMobEffect;
     }