+++ 1.21.8/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.8.1-1.21.8] - 2025-07-26
+
+### Added
+
+- Add a decorator string before every effect description, which also allows for setting an indent and for controlling the description formatting via the use of formatting codes
+
+### Changed
+
+- When a shift key has to be held to view effect descriptions, a hint is now included on the tooltip
+
+### Fixed
+
+- Fix a crash that would happen when trying to gather descriptions for an item that has the same effect multiple times
+
 ## [v21.8.0-1.21.8] - 2025-07-18
+
 - Update to Minecraft 1.21.8